% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Functions to Analyze Data Set on Earthquake Activity for Capstone Proj in Coursera.
Read in the tab delimited data file from NOAA.
Creates a data frame from raw data file.
data obtained from (National Geophysical Data Center / World Data Service (NGDC/WDS)
                          Significant Earthquake Database. National Geophysical Data Center, NOAA. 
                          doi10.7289/V5TD9V7K)
Cleaning the data by formatting the date time & Latitude & Longitude}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{is the data frame with data and NA from previous step.}
}
\value{
this function will return 'cleaned' data for the previously created data frame
}
\description{
Functions to Analyze Data Set on Earthquake Activity for Capstone Proj in Coursera.
Read in the tab delimited data file from NOAA.
Creates a data frame from raw data file.
data obtained from (National Geophysical Data Center / World Data Service (NGDC/WDS)
                          Significant Earthquake Database. National Geophysical Data Center, NOAA. 
                          doi10.7289/V5TD9V7K)
Cleaning the data by formatting the date time & Latitude & Longitude
}
\examples{
\dontrun{
df = readr::read_delim("./inst/extdata/results.tsv", delim = "\\t") \%>\% eq_clean_data()
}
}
