% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Make a label for the popups
Furthering the detailed information in the annotation boxes}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{the cleaned and fileterd data frame}
}
\value{
label information
}
\description{
Make a label for the popups
Furthering the detailed information in the annotation boxes
}
