% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location name}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{is the data frame from intial cleaning of the raw data}
}
\description{
This will clean the data to remove colon from Location Names and convert it to title case for later use.
}
\examples{
\dontrun{
df = readr::read_delim("./inst/extdata/results.tsv", delim = "\\t") \%>\% eq_clean_data() \%>\% 
eq_location_clean()
}
}
