% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_map}
\alias{eq_map}
\title{Plot the map}
\usage{
eq_map(data, annot_col = "popup_text")
}
\arguments{
\item{data}{the cleaned and filtered data frame}

\item{annot_col}{the information used in popup label on map}
}
\value{
this will return a map with points plotted for records in the US.
}
\description{
Plots the global map with annotations for earthquakes within filtered data set.
}
\examples{
\dontrun{
df = readr::read_delim("./inst/extdata/results.tsv", delim = "\\t") \%>\% eq_clean_data() \%>\% eq_location_clean()
geom_timeline(df)
eq_create_label(df)
eq_map(df)}
}
