% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_readIn_data}
\alias{eq_readIn_data}
\title{Functions to Analyze Data Set for earthquakes for Capstone Proj in Coursera.
Read in the tab delimited data file from NOAA.
Creates a data frame from raw data file.}
\usage{
eq_readIn_data(filename)
}
\arguments{
\item{filename}{the name of the .tsv file}
}
\value{
this funtion will return data frame from tab separated file and fill balnks with NA.
}
\description{
Functions to Analyze Data Set for earthquakes for Capstone Proj in Coursera.
Read in the tab delimited data file from NOAA.
Creates a data frame from raw data file.
}
\note{
readr read.delim will read the file and creates a data frame and must have an element in each column,
therefore added the \code{fill=TRUE} to fill in the blanks that are present in table with NA's
}
\examples{
\dontrun {
eq_readIn_data(filename="results.tsv")
}
}
