% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline labels
Add text labels for each plotted dot on timeline}
\usage{
geom_timeline_label(data)
}
\arguments{
\item{data}{the cleaned data}
}
\value{
this will return a timeline as above but with labels for data points
}
\description{
Timeline labels
Add text labels for each plotted dot on timeline
}
\examples{
 \dontrun {
geom_timeline_label(df_filtered)
}
}
