% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning of the data.
This will clean the data to remove colon from Location Names and convert it to title case for later use.}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{is the data frame with data and NA from previous step}
}
\value{
this function will return 'cleaned' data for the previously created data frame
}
\description{
Cleaning of the data.
This will clean the data to remove colon from Location Names and convert it to title case for later use.
}
\note{
This function will clean up original LOCATION_NAME column data to remove colon and convert to title case for use later.
}
\examples{
\dontrun {
eq_clean_data(df)
}
}
