% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Make a label for the popups}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{the cleaned and fileterd data frame}
}
\value{
label information

subdf an updated data frame with label information
}
\description{
Furthering the detailed information in the annotation boxes
}
\examples{
\dontrun{
df = readr::read_delim("./inst/extdata/results.tsv", delim = "\\t") \%>\% eq_clean_data() \%>\% 
eq_location_clean()
geom_timeline_label(df)
eq_create_label(df)}
}
