% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{eq_map}
\alias{eq_map}
\title{Plot the map
PLots the global map with annotations for earthquakes within filtered data set.}
\usage{
eq_map(data, annot_col = "YEAR")
}
\arguments{
\item{data}{the cleaned and filtered data frame}

\item{annot_col}{the information used in popup label on map}
}
\value{
this will return a map with points plotted for records in the US.
}
\description{
Plot the map
PLots the global map with annotations for earthquakes within filtered data set.
}
\examples{
\dontrun {
eq_map(df_filtered)
}
}
