% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaker.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Further cleaning of the data.
Filters the data by date and country.}
\usage{
geom_timeline(data)
}
\arguments{
\item{data}{the cleaned data to load into function}
}
\value{
this function will return a timeline showing dots of relative size color relating to data
}
\description{
Further cleaning of the data.
Filters the data by date and country.
}
\note{
can include more than one country name to display multiple timelines.
}
\examples{
 \dontrun {
geom_timeline(df_clean)
}
}
